Create a test using Iterator instead of Stream.
I wanted to create a test that used as little memory as possible
outside of the async parser. So I decided to use Iterator[String]
and a while loop. It's ugly but it works.

I also removed the per-element string concatenation in favor of
including the comma in the data definition.

My test with these changes showed no memory growth. I managed to
quickly get to 6M element parses without any full GCs, memory
growth, etc.